+++ app/src/main/java/net/programmierecke/radiodroid2/service/PlayerServiceUtil.java
@@ -186,6 +186,48 @@ public class PlayerServiceUtil {
         }
     }
 
+    public static void seekBackward(long ms) {
+        if (itsPlayerService != null) {
+            try {
+                itsPlayerService.SeekBackward(ms);
+            } catch (RemoteException e) {
+                Log.e("", "" + e);
+            }
+        }
+    }
+
+    public static boolean canSeekBackward() {
+        if (itsPlayerService != null) {
+            try {
+                return itsPlayerService.canSeekBackward();
+            } catch (RemoteException e) {
+                Log.e("", "" + e);
+            }
+        }
+        return false;
+    }
+
+    public static void seekToLive() {
+        if (itsPlayerService != null) {
+            try {
+                itsPlayerService.SeekToLive();
+            } catch (RemoteException e) {
+                Log.e("", "" + e);
+            }
+        }
+    }
+
+    public static boolean isBehindLive() {
+        if (itsPlayerService != null) {
+            try {
+                return itsPlayerService.isBehindLive();
+            } catch (RemoteException e) {
+                Log.e("", "" + e);
+            }
+        }
+        return false;
+    }
+
     public static void pause(PauseReason pauseReason) {
         if (itsPlayerService != null) {
             try {